New method added to run machine over array of messages
I realized that char ** will not do as message arrays. Because we
cannot know the length of the message beforehand. This means that
the FSM will probably fail when any state points to more than one
state.

So I need to change the char ** in the NFA struct with array of
message_structs